extract communication to a separate folder, method class wip
