+++ src/communication/messageIds.ts
@@ -8,4 +8,5 @@ export const INTERFACE_MESSAGES = {
 export const SDK_MESSAGES = {
   SAFE_APP_SDK_INITIALIZED: 'SAFE_APP_SDK_INITIALIZED',
   SEND_TRANSACTIONS: 'SEND_TRANSACTIONS',
+  RPC_CALL: 'RPC_CALL',
 } as const;
